+++ src/pipecat/services/groq/tts.py
@@ -10,7 +10,7 @@ from loguru import logger
 from pydantic import BaseModel
 
 from pipecat.frames.frames import Frame, TTSAudioRawFrame, TTSStartedFrame, TTSStoppedFrame
-from pipecat.services.ai_services import TTSService
+from pipecat.services.tts_service import TTSService
 from pipecat.transcriptions.language import Language
 
 try: